Visualization for explanatory data analysis - TMT experiment

To illustrate the quantitative data and quality control of MS runs, dataProcessPlotsTMT takes the quantitative data from converter functions (PDtoMSstatsTMTFormat, MaxQtoMSstatsTMTFormat, SpectroMinetoMSstatsTMTFormat) as input and generate two types of figures in pdf files as output : (1) profile plot (specify "ProfilePlot" in option type), to identify the potential sources of variation for each protein; (2) quality control plot (specify "QCPlot" in option type), to evaluate the systematic bias between MS runs.

dataProcessPlotsTMT(
  data.peptide,
  data.summarization,
  type,
  ylimUp = FALSE,
  ylimDown = FALSE,
  x.axis.size = 10,
  y.axis.size = 10,
  text.size = 4,
  text.angle = 90,
  legend.size = 7,
  dot.size.profile = 2,
  ncol.guide = 5,
  width = 10,
  height = 10,
  which.Protein = "all",
  originalPlot = TRUE,
  summaryPlot = TRUE,
  address = ""
)

Arguments

data.peptide
name of the data with peptide level, which can be the output of converter functions(PDtoMSstatsTMTFormat, MaxQtoMSstatsTMTFormat, SpectroMinetoMSstatsTMTFormat).

data.summarization
name of the data with protein-level, which can be the output of proteinSummarization function.

type
choice of visualization. "ProfilePlot" represents profile plot of log intensities across MS runs. "QCPlot" represents box plots of log intensities across channels and MS runs.

ylimUp
upper limit for y-axis in the log scale. FALSE(Default) for Profile Plot and QC Plot uses the upper limit as rounded off maximum of log2(intensities) after normalization + 3..

ylimDown
lower limit for y-axis in the log scale. FALSE(Default) for Profile Plot and QC Plot uses 0..

x.axis.size
size of x-axis labeling for "Run" and "channel in Profile Plot and QC Plot.

y.axis.size
size of y-axis labels. Default is 10.

text.size
size of labels represented each condition at the top of Profile plot and QC plot. Default is 4.

text.angle
angle of labels represented each condition at the top of Profile plot and QC plot. Default is 0.

legend.size
size of legend above Profile plot. Default is 7.

dot.size.profile
size of dots in Profile plot. Default is 2.

ncol.guide
number of columns for legends at the top of plot. Default is 5.

width
width of the saved pdf file. Default is 10.

height
height of the saved pdf file. Default is 10.

which.Protein
Protein list to draw plots. List can be names of Proteins or order numbers of Proteins. Default is "all", which generates all plots for each protein. For QC plot, "allonly" will generate one QC plot with all proteins.

originalPlot
TRUE(default) draws original profile plots, without normalization.

summaryPlot
TRUE(default) draws profile plots with protein summarization for each channel and MS run.

address
the name of folder that will store the results. Default folder is the current working directory. The other assigned folder has to be existed under the current working directory. An output pdf file is automatically created with the default name of "ProfilePlot.pdf" or "QCplot.pdf". The command address can help to specify where to store the file as well as how to modify the beginning of the file name. If address=FALSE, plot will be not saved as pdf file but showed in window.

Value

plot or pdf

Examples

data(input.pd)
quant.msstats <- proteinSummarization(input.pd,
                                      method="msstats",
                                      global_norm=TRUE,
                                      reference_norm=TRUE)

#> Joining, by = c("Run", "Channel")

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture1_01.raw ( 1  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     4-29
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [K].sTPSGFTLDDVIQTGVDNPGHPYIMTVGcVAGDEESYEVFk.[D]_4_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture1_02.raw ( 2  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-33
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [R].eVLGDAVPDEILIEAVLk.[N]_3_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture1_03.raw ( 3  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-29
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [K].qQQDQVDr.[N]_2_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture2_01.raw ( 4  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     1-28
#> # of Transitions/Peptide   1-1

#> 
#> ** 1 Proteins have only single transition : Consider excluding this protein from the dataset. (Q9Y450) 

#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [K].dYEFMWNPHLGYILTcPSNLGTGLr.[A]_3_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture2_02.raw ( 5  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     1-30
#> # of Transitions/Peptide   1-1

#> 
#> ** 1 Proteins have only single transition : Consider excluding this protein from the dataset. (Q9Y450) 

#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 2

#>     -> [K].qQQDQVDr.[N]_2_NA_NA, [R].nLPQYVSNELLEEAFSVFGQVEr.[A]_3_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture2_03.raw ( 6  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     2-30
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 0

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture3_01.raw ( 7  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     4-31
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 0

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture3_02.raw ( 8  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-30
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [K].vDIVAINDPFIDLNYMVYMFQYDSTHGk.[F]_3_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture3_03.raw ( 9  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     5-30
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 3

#>     -> [K].vDIVAINDPFIDLNYMVYMFQYDSTHGk.[F]_3_NA_NA, [R].iPSAVGYQPTLATDMGTMQEr.[I]_2_NA_NA, [R].gAMPPAPVPAGTPAPPGPATMMPDGTLGLTPPTTEr.[F]_4_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture4_01.raw ( 10  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-31
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 2

#>     -> [K].sTPSGFTLDDVIQTGVDNPGHPYIMTVGcVAGDEESYEVFk.[D]_4_NA_NA, [K].qQQDQVDr.[N]_2_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture4_02.raw ( 11  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-31
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [R].fcTGLTQIETLFk.[S]_2_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture4_03.raw ( 12  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     1-31
#> # of Transitions/Peptide   1-1

#> 
#> ** 1 Proteins have only single transition : Consider excluding this protein from the dataset. (Q9Y450) 

#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 0

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture5_01.raw ( 13  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-34
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 0

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture5_02.raw ( 14  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     2-30
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 3

#>     -> [R].mGQMAMGGAmGINNr.[G]_2_NA_NA, [R].nLPQYVSNELLEEAFSVFGQVER.[A]_3_NA_NA, [R].dQNAEQIr.[L]_2_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture5_03.raw ( 15  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     5-32
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 4

#>     -> [K].gFQQILAGEYDHLPEQAFYmVGPIEEAVAk.[A]_3_NA_NA, [K].qFAPIHAEAPEFMEMSVEQEILVTGIk.[V]_4_NA_NA, [K].qQQDQVDr.[N]_2_NA_NA, [R].dQNAEQIr.[L]_2_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Normalization between MS runs for Protein : P04406 ( 1  of  10 )

#> Normalization between MS runs for Protein : P06576 ( 2  of  10 )

#> Normalization between MS runs for Protein : P12277 ( 3  of  10 )

#> Normalization between MS runs for Protein : P23919 ( 4  of  10 )

#> Normalization between MS runs for Protein : P31947 ( 5  of  10 )

#> Normalization between MS runs for Protein : Q15233 ( 6  of  10 )

#> Normalization between MS runs for Protein : Q16181 ( 7  of  10 )

#> Normalization between MS runs for Protein : Q9NSD9 ( 8  of  10 )

#> Normalization between MS runs for Protein : Q9UGP8 ( 9  of  10 )

#> Normalization between MS runs for Protein : Q9Y450 ( 10  of  10 )

## Profile plot
dataProcessPlotsTMT(data.peptide=input.pd,
                   data.summarization=quant.msstats,
                   type='ProfilePlot',
                   width = 21,
                   height = 7)

#> Warning: Removed 16 rows containing missing values (geom_point).

#> Drew the Profile plot for  P04406 ( 1  of  10 )

#> Warning: Removed 29 rows containing missing values (geom_point).

#> Warning: Removed 1 row(s) containing missing values (geom_path).

#> Drew the Profile plot for  P06576 ( 2  of  10 )

#> Warning: Removed 23 rows containing missing values (geom_point).

#> Warning: Removed 1 row(s) containing missing values (geom_path).

#> Drew the Profile plot for  P12277 ( 3  of  10 )

#> Warning: Removed 2 rows containing missing values (geom_point).

#> Drew the Profile plot for  P23919 ( 4  of  10 )

#> Drew the Profile plot for  P31947 ( 5  of  10 )

#> Warning: Removed 52 rows containing missing values (geom_point).

#> Warning: Removed 3 row(s) containing missing values (geom_path).

#> Drew the Profile plot for  Q15233 ( 6  of  10 )

#> Warning: Removed 2 rows containing missing values (geom_point).

#> Drew the Profile plot for  Q16181 ( 7  of  10 )

#> Warning: Removed 2 rows containing missing values (geom_point).

#> Drew the Profile plot for  Q9NSD9 ( 8  of  10 )

#> Warning: Removed 8 rows containing missing values (geom_point).

#> Drew the Profile plot for  Q9UGP8 ( 9  of  10 )

#> Warning: Removed 6 rows containing missing values (geom_point).

#> Drew the Profile plot for  Q9Y450 ( 10  of  10 )

#> Warning: Removed 16 rows containing missing values (geom_point).

#> Warning: Removed 16 rows containing missing values (geom_point).

#> Drew the Profile plot with summarization for  P04406 ( 1  of  10 )

#> Warning: Removed 29 rows containing missing values (geom_point).

#> Warning: Removed 1 row(s) containing missing values (geom_path).

#> Warning: Removed 29 rows containing missing values (geom_point).

#> Drew the Profile plot with summarization for  P06576 ( 2  of  10 )

#> Warning: Removed 23 rows containing missing values (geom_point).

#> Warning: Removed 1 row(s) containing missing values (geom_path).

#> Warning: Removed 23 rows containing missing values (geom_point).

#> Drew the Profile plot with summarization for  P12277 ( 3  of  10 )

#> Warning: Removed 2 rows containing missing values (geom_point).

#> Warning: Removed 2 rows containing missing values (geom_point).

#> Drew the Profile plot with summarization for  P23919 ( 4  of  10 )

#> Drew the Profile plot with summarization for  P31947 ( 5  of  10 )

#> Warning: Removed 52 rows containing missing values (geom_point).

#> Warning: Removed 3 row(s) containing missing values (geom_path).

#> Warning: Removed 52 rows containing missing values (geom_point).

#> Drew the Profile plot with summarization for  Q15233 ( 6  of  10 )

#> Warning: Removed 2 rows containing missing values (geom_point).

#> Warning: Removed 2 rows containing missing values (geom_point).

#> Drew the Profile plot with summarization for  Q16181 ( 7  of  10 )

#> Warning: Removed 2 rows containing missing values (geom_point).

#> Warning: Removed 2 rows containing missing values (geom_point).

#> Drew the Profile plot with summarization for  Q9NSD9 ( 8  of  10 )

#> Warning: Removed 8 rows containing missing values (geom_point).

#> Warning: Removed 8 rows containing missing values (geom_point).

#> Drew the Profile plot with summarization for  Q9UGP8 ( 9  of  10 )

#> Warning: Removed 6 rows containing missing values (geom_point).

#> Warning: Removed 6 rows containing missing values (geom_point).

#> Drew the Profile plot with summarization for  Q9Y450 ( 10  of  10 )

## NottoRun: QC plot
# dataProcessPlotsTMT(data.peptide=input.pd,
                    # data.summarization=quant.msstats,
                    # type='QCPlot',
                    # width = 21,
                    # height = 7)