Finding differentially abundant proteins across conditions in TMT experiment

Tests for significant changes in protein abundance across conditions based on a family of linear mixed-effects models in TMT experiment. Experimental design of case-control study (patients are not repeatedly measured) is automatically determined based on proper statistical model.

groupComparisonTMT(
  data,
  contrast.matrix = "pairwise",
  moderated = FALSE,
  adj.method = "BH",
  remove_norm_channel = TRUE,
  remove_empty_channel = TRUE
)

Arguments

data
Name of the output of proteinSummarization function. It should have columns named Protein, Mixture, TechRepMixture, Run, Channel, Condition, BioReplicate, Abundance.

contrast.matrix
Comparison between conditions of interests. 1) default is 'pairwise', which compare all possible pairs between two conditions. 2) Otherwise, users can specify the comparisons of interest. Based on the levels of conditions, specify 1 or -1 to the conditions of interests and 0 otherwise. The levels of conditions are sorted alphabetically.

moderated
TRUE will moderate t statistic; FALSE (default) uses ordinary t statistic.

adj.method
adjusted method for multiple comparison. "BH" is default.

remove_norm_channel
TRUE(default) removes 'Norm' channels from protein level data.

remove_empty_channel
TRUE(default) removes 'Empty' channels from protein level data.

Value

data.frame with result of inference

Examples

data(input.pd)
# use protein.summarization() to get protein abundance data
quant.pd.msstats <- proteinSummarization(input.pd,
                                       method="msstats",
                                       global_norm=TRUE,
                                       reference_norm=TRUE)

#> Joining, by = c("Run", "Channel")

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture1_01.raw ( 1  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     4-29
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [K].sTPSGFTLDDVIQTGVDNPGHPYIMTVGcVAGDEESYEVFk.[D]_4_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture1_02.raw ( 2  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-33
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [R].eVLGDAVPDEILIEAVLk.[N]_3_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture1_03.raw ( 3  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-29
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [K].qQQDQVDr.[N]_2_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture2_01.raw ( 4  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     1-28
#> # of Transitions/Peptide   1-1

#> 
#> ** 1 Proteins have only single transition : Consider excluding this protein from the dataset. (Q9Y450) 

#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [K].dYEFMWNPHLGYILTcPSNLGTGLr.[A]_3_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture2_02.raw ( 5  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     1-30
#> # of Transitions/Peptide   1-1

#> 
#> ** 1 Proteins have only single transition : Consider excluding this protein from the dataset. (Q9Y450) 

#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 2

#>     -> [K].qQQDQVDr.[N]_2_NA_NA, [R].nLPQYVSNELLEEAFSVFGQVEr.[A]_3_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture2_03.raw ( 6  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     2-30
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 0

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture3_01.raw ( 7  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     4-31
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 0

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture3_02.raw ( 8  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-30
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [K].vDIVAINDPFIDLNYMVYMFQYDSTHGk.[F]_3_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture3_03.raw ( 9  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     5-30
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 3

#>     -> [K].vDIVAINDPFIDLNYMVYMFQYDSTHGk.[F]_3_NA_NA, [R].iPSAVGYQPTLATDMGTMQEr.[I]_2_NA_NA, [R].gAMPPAPVPAGTPAPPGPATMMPDGTLGLTPPTTEr.[F]_4_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture4_01.raw ( 10  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-31
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 2

#>     -> [K].sTPSGFTLDDVIQTGVDNPGHPYIMTVGcVAGDEESYEVFk.[D]_4_NA_NA, [K].qQQDQVDr.[N]_2_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture4_02.raw ( 11  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-31
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 1

#>     -> [R].fcTGLTQIETLFk.[S]_2_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture4_03.raw ( 12  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     1-31
#> # of Transitions/Peptide   1-1

#> 
#> ** 1 Proteins have only single transition : Consider excluding this protein from the dataset. (Q9Y450) 

#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 0

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture5_01.raw ( 13  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     3-34
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 0

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture5_02.raw ( 14  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     2-30
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 3

#>     -> [R].mGQMAMGGAmGINNr.[G]_2_NA_NA, [R].nLPQYVSNELLEEAFSVFGQVER.[A]_3_NA_NA, [R].dQNAEQIr.[L]_2_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Summarizing for Run : 161117_SILAC_HeLa_UPS1_TMT10_Mixture5_03.raw ( 15  of  15 )

#> ** Use all features that the dataset origianally has.

#>                        
#>   Summary of Features :
#>                          count
#> # of Protein                10
#> # of Peptides/Protein     5-32
#> # of Transitions/Peptide   1-1
#>                       
#>   Summary of Samples :
#>                            0.125 0.5 0.667 1 Norm
#> # of MS runs                   2   2     2 2    2
#> # of Biological Replicates     1   1     1 1    1
#> # of Technical Replicates      2   2     2 2    2

#> 
#>  Summary of Missingness :

#>   # transitions are completely missing in at least one of the conditions : 4

#>     -> [K].gFQQILAGEYDHLPEQAFYmVGPIEEAVAk.[A]_3_NA_NA, [K].qFAPIHAEAPEFMEMSVEQEILVTGIk.[V]_4_NA_NA, [K].qQQDQVDr.[N]_2_NA_NA, [R].dQNAEQIr.[L]_2_NA_NA ...

#> 
#>   # run with 75% missing observations: 0

#> 
#>  == Start the summarization per subplot...

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> 
#>  == the summarization per subplot is done.

#> ** Protein-level summarization done by MSstats.

#> Normalization between MS runs for Protein : P04406 ( 1  of  10 )

#> Normalization between MS runs for Protein : P06576 ( 2  of  10 )

#> Normalization between MS runs for Protein : P12277 ( 3  of  10 )

#> Normalization between MS runs for Protein : P23919 ( 4  of  10 )

#> Normalization between MS runs for Protein : P31947 ( 5  of  10 )

#> Normalization between MS runs for Protein : Q15233 ( 6  of  10 )

#> Normalization between MS runs for Protein : Q16181 ( 7  of  10 )

#> Normalization between MS runs for Protein : Q9NSD9 ( 8  of  10 )

#> Normalization between MS runs for Protein : Q9UGP8 ( 9  of  10 )

#> Normalization between MS runs for Protein : Q9Y450 ( 10  of  10 )

test.pairwise <- groupComparisonTMT(quant.pd.msstats, moderated = TRUE)

#> Model fitting for 10 proteins:

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> Testing for 10 proteins:

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

# Only compare condition 0.125 and 1
levels(quant.pd.msstats$Condition)

#> [1] "0.125" "0.5"   "0.667" "1"    

# Compare condition 1 and 0.125
comparison<-matrix(c(-1,0,0,1),nrow=1)

# Set the nafmes of each row
row.names(comparison)<-"1-0.125"

# Set the column names
colnames(comparison)<- c("0.125", "0.5", "0.667", "1")
test.contrast <- groupComparisonTMT(data = quant.pd.msstats,
contrast.matrix = comparison,
moderated = TRUE)

#> Model fitting for 10 proteins:

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%

#> Testing for 10 proteins:

#>   |                                                                              |                                                                      |   0%  |                                                                              |=======                                                               |  10%  |                                                                              |==============                                                        |  20%  |                                                                              |=====================                                                 |  30%  |                                                                              |============================                                          |  40%  |                                                                              |===================================                                   |  50%  |                                                                              |==========================================                            |  60%  |                                                                              |=================================================                     |  70%  |                                                                              |========================================================              |  80%  |                                                                              |===============================================================       |  90%  |                                                                              |======================================================================| 100%